First Commit
This is my more advanced game that I created using module techniques.

I discovered in the original that a single bug could cause my entire game to fail.. (obviously not good). Furthermore, I was unable to scale it up to this level because it didn't adhere to RESTful programming.

This was my solution a vast collection of module scripts that allowed me to implement VASTLY more code without causing undue stress on either the server nor client, while making code more re-usable.